Misc code refinements

+++ Autocycler-Rust/src/unitig_graph.rs
@@ -397,12 +397,12 @@ impl UnitigGraph {
         // Given a path (vector of unitig IDs and strands), this function returns the sequence
         // traced by that path. It also requires a unitig index so it can quickly look up unitigs
         // by their number.
-        let half_k = self.k_size / 2;
+        let half_k = (self.k_size / 2) as usize;
         let mut sequence = Vec::new();
         for (i, (unitig_num, strand)) in path.iter().enumerate() {
             let unitig = self.unitig_index.get(unitig_num).unwrap();
-            let upstream = if i == 0 { half_k as usize } else { 0 };
-            let downstream = if i == path.len() - 1 { half_k as usize } else { 0 };
+            let upstream = if i == 0 { half_k } else { 0 };
+            let downstream = if i == path.len() - 1 { half_k } else { 0 };
             sequence.push(String::from_utf8(unitig.borrow().get_seq(*strand, upstream, downstream)).unwrap());
         }
         sequence.into_iter().collect()
